+viz directory; +tools test/visualization test; +renamed plot_cases_by_location & + new past_num_days arg
