+++ tests/test_outbreak_tools.py
@@ -1,26 +1,10 @@
 from outbreak_tools import outbreak_tools
-import altair as alt
+import pytest
+import pandas as pd
 
+def _test_lookup_table():
+    output = outbreak_tools.id_lookup([''], table=True)
+    assert isinstance(output, pd.core.frame.DataFrame), 'When table is true should return dataframe'
 
-def _test_single_location():
-    location = 'USA_US-CA'
-    expected_loc_name = 'California'
-    plot = outbreak_tools.plot_case_increase(location)
-    assert type(plot) == alt.vegalite.v4.api.Chart, 'Visualization not returning correct format'
-    plotted_loc = plot.data.admin1.unique()[0]
-    assert plotted_loc == expected_loc_name, 'Visualization not plotting correct location'
-
-def _test_multiple_locations():
-    locations = ['USA_US-CA', 'USA_US-OR', 'USA_US-WA', 'USA_US-NV']
-    plot = outbreak_tools.plot_case_increase(locations)
-    assert type(plot) == alt.vegalite.v4.api.Chart, 'Visualization not returning correct format'
-    plotted_locs = plot.data.admin1.unique()
-    assert len(plotted_locs) == len(locations), 'Visualization not plotting all locations'
-
-def test_plot_case_increase():
-    """
-    Tests the cases_by_location visualization tool in outbreak_tools
-    """
-    _test_single_location()
-    _test_multiple_locations()
-
+def test_id_lookup():
+    _test_lookup_table()
